*Double swap done on unint64_t for higher compatibility. *Packets list not backtracked anymore.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/wsim@361 c0ef0dea-404d-0410-ad65-84cceb5f200a

+++ utils/wsnet1/include/worldsens.h
@@ -143,7 +143,9 @@ struct PACKED _worldsens_c_tx_pkt
   uint64_t duration;
   uint32_t frequency;
   int	   modulation;
-  double   tx_mW;
+/* To be more portable double value are stored into uint64_t variables to be swap and unswap */
+/* To get their floating values, a cast pointer must be used : (double *) xxx */
+  uint64_t tx_mW;       /* double */
   int	   pkt_seq;
   char	   data;
 };
@@ -269,8 +271,10 @@ struct PACKED _worldsens_data
 {
   int		     node;
   char		     data;
-  double	     rx_mW;
-  double	     SiNR;
+/* To be more portable double value are stored into uint64_t variables to be swap and unswap */
+/* To get their floating values, a cast pointer must be used : (double *) xxx */
+  uint64_t	     rx_mW; /* double */
+  uint64_t	     SiNR;  /* double */
 };
